김기현 - 그룹 변경 및 생성, 각 그룹 최근 메세지 출력 테스트 구현

+++ src/main/java/com/sist/web/chat/group/controller/GroupChatController.java
@@ -4,6 +4,7 @@ import org.springframework.messaging.handler.annotation.DestinationVariable;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import com.sist.web.chat.group.service.*;
 import com.sist.web.chat.group.vo.*;
@@ -23,7 +24,7 @@ public class GroupChatController {
 	}
 	
 	@MessageMapping("/chats/groups/{groupId}")
-	public void sendGroupChat(@DestinationVariable Long groupId, GroupChatVO vo) {
+	public void sendGroupChat(@DestinationVariable int groupId, GroupChatVO vo) {
 		String senderId = "user";
 		System.out.println("사용자: " + senderId + " / 메시지: " + vo.getContent());
